services/connectors/miniflux/pyproject.toml
[project]
name = "connector-miniflux"
version = "0.1.0"
description = "Miniflux connector"
requires-python = ">=3.12"
dependencies = [
  "httpx>=0.27.0",
  "python-dateutil>=2.9.0",
]

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/connector_miniflux"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
